+++ src/init.c
@@ -5,6 +5,9 @@
 
 static const R_CallMethodDef callMethods[]  = {
   {"keypress", (DL_FUNC) &keypress, 1},
+  {"save_term_status", (DL_FUNC) &save_term_status, 0},
+  {"restore_term_status", (DL_FUNC) &restore_term_status, 0},
+  {"set_term_echo", (DL_FUNC) &set_term_echo, 1},
   {NULL, NULL, 0}
 };
 
@@ -15,4 +18,8 @@ void R_init_keypress(DllInfo *dll) {
 
   R_RegisterCCallable("keypress", "keypress", (DL_FUNC) keypress);
   R_RegisterCCallable("keypress_read", "keypress_read", (DL_FUNC) keypress_read);
-}
+  R_RegisterCCallable("save_term_status", "save_term_status", (DL_FUNC) save_term_status);
+  R_RegisterCCallable("restore_term_status", "restore_term_status", (DL_FUNC) restore_term_status);
+  R_RegisterCCallable("set_term_echo", "set_term_echo", (DL_FUNC) set_term_echo);
+
+  }