+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # development version
 
+- Support non-blocking press on Windows (Rterm)
+
 # 1.2.0
 
 - Support RStudio terminal.