+++ NEWS.md
@@ -1,8 +1,8 @@
-# more development (wes)
-
-- Allow turning echo off for non-blocking linux/OSX terminals
 
 # development version
+
+- Allow turning echo off for non-blocking Linux and macOS terminals
+  (#11, @weshinsley)
 
 # 1.2.0