+++ NEWS.md
@@ -1,5 +1,4 @@
-
-# development version
+# keypress 1.3.0
 
 - Allow turning echo off for non-blocking Linux and macOS terminals
   (#11, @weshinsley)